logrotate: Allow end users to configure frequency and rotation.
This commit added frequency and rotate params to nginxpack::logrotate
class. Default values are changed from daily (frequency) to weekly, and
10 day (rotate) to 52 weeks. (Due to some legal constraints…).

The kill signal was also updated according Debian package¹, nginx
support a custom signal on USR1, HUP is a bit not safe because it lead to
service distribution (during the SIGHUP).

¹http://anonscm.debian.org/cgit/collab-maint/nginx.git/plain/debian/nginx-common.nginx.logrotate